requirements.txt
numpy>=1.21.0
python-dotenv>=0.19.0
matplotlib>=3.5.0
